Extract agent run folder preparation to a new module and rename for clarity
