+++ agent/src/cli.rs
@@ -14,17 +14,10 @@
 
 use regex::Regex;
 
-use std::path::Path;
-
-use crate::io_utils::FileSystemError;
-#[cfg_attr(test, mockall_double::double)]
-use crate::io_utils::{filesystem, Directory};
+use crate::io_utils::DEFAULT_RUN_FOLDER;
 use clap::Parser;
 use common::objects::STR_RE_AGENT;
 use common::DEFAULT_SERVER_ADDRESS;
-
-const DEFAULT_RUN_FOLDER: &str = "/tmp/ankaios/";
-const RUNFOLDER_SUFFIX: &str = "_io";
 
 // [impl->swdd~agent-naming-convention~1]
 fn validate_agent_name(name: &str) -> Result<String, String> {
@@ -76,27 +69,6 @@ pub struct Arguments {
     pub key_pem: Option<String>,
 }
 
-impl Arguments {
-    // [impl->swdd~agent-prepares-dedicated-run-folder~1]
-    pub fn get_agent_run_directory(&self) -> Result<Directory, FileSystemError> {
-        let base_path = Path::new(&self.run_folder);
-        let agent_run_folder = base_path.join(format!("{}{}", self.agent_name, RUNFOLDER_SUFFIX));
-
-        // If the default base dir is used, we need to take care of its creation
-        if !filesystem::exists(base_path) {
-            if Some(DEFAULT_RUN_FOLDER) == base_path.to_str() {
-                filesystem::make_dir(base_path)?;
-
-                filesystem::set_permissions(base_path, 0o777)?;
-            } else {
-                return Err(FileSystemError::NotFoundDirectory(base_path.into()));
-            }
-        }
-
-        Directory::new(agent_run_folder)
-    }
-}
-
 pub fn parse() -> Arguments {
     Arguments::parse()
 }
@@ -111,12 +83,6 @@ pub fn parse() -> Arguments {
 
 #[cfg(test)]
 mod tests {
-    use super::{Arguments, FileSystemError, Path, DEFAULT_RUN_FOLDER};
-    use crate::io_utils::generate_test_directory_mock;
-    use crate::io_utils::mock_filesystem;
-    use common::DEFAULT_SERVER_ADDRESS;
-
-    use mockall::predicate;
 
     // [utest->swdd~agent-naming-convention~1]
     #[test]
@@ -133,199 +99,5 @@ mod tests {
         assert!(super::validate_agent_name("a.b").is_err());
         assert!(super::validate_agent_name("a_b_%#").is_err());
         assert!(super::validate_agent_name("a b").is_err());
-    }
-
-    // [utest->swdd~agent-prepares-dedicated-run-folder~1]
-    #[test]
-    fn utest_arguments_get_run_directory_use_default_directory_create() {
-        let _guard = crate::test_helper::MOCKALL_CONTEXT_SYNC.get_lock();
-
-        let args = Arguments {
-            agent_name: "test_agent_name".to_owned(),
-            server_url: DEFAULT_SERVER_ADDRESS.to_string(),
-            run_folder: DEFAULT_RUN_FOLDER.to_owned(),
-            insecure: true,
-            ca_pem: None,
-            crt_pem: None,
-            key_pem: None,
-        };
-
-        let exists_mock_context = mock_filesystem::exists_context();
-        exists_mock_context
-            .expect()
-            .with(predicate::eq(Path::new(DEFAULT_RUN_FOLDER).to_path_buf()))
-            .return_const(false);
-
-        let mk_dir_context = mock_filesystem::make_dir_context();
-        mk_dir_context
-            .expect()
-            .with(predicate::eq(Path::new(DEFAULT_RUN_FOLDER).to_path_buf()))
-            .return_once(|_| Ok(()));
-
-        let set_permissions_mock_context = mock_filesystem::set_permissions_context();
-        set_permissions_mock_context
-            .expect()
-            .with(
-                predicate::eq(Path::new(DEFAULT_RUN_FOLDER).to_path_buf()),
-                predicate::eq(0o777),
-            )
-            .return_once(|_, _| Ok(()));
-
-        let _directory_mock_context =
-            generate_test_directory_mock(DEFAULT_RUN_FOLDER, "test_agent_name_io");
-
-        assert!(args.get_agent_run_directory().is_ok());
-    }
-
-    // [utest->swdd~agent-prepares-dedicated-run-folder~1]
-    #[test]
-    fn utest_arguments_get_run_directory_use_default_directory_create_fails() {
-        let _guard = crate::test_helper::MOCKALL_CONTEXT_SYNC.get_lock();
-
-        let args = Arguments {
-            agent_name: "test_agent_name".to_owned(),
-            server_url: DEFAULT_SERVER_ADDRESS.to_string(),
-            run_folder: DEFAULT_RUN_FOLDER.to_owned(),
-            insecure: true,
-            ca_pem: None,
-            crt_pem: None,
-            key_pem: None,
-        };
-
-        let exists_mock_context = mock_filesystem::exists_context();
-        exists_mock_context
-            .expect()
-            .with(predicate::eq(Path::new(DEFAULT_RUN_FOLDER).to_path_buf()))
-            .return_const(false);
-
-        let mk_dir_context = mock_filesystem::make_dir_context();
-        mk_dir_context
-            .expect()
-            .with(predicate::eq(Path::new(DEFAULT_RUN_FOLDER).to_path_buf()))
-            .return_once(|_| {
-                Err(FileSystemError::CreateDirectory(
-                    Path::new("/tmp/x").as_os_str().to_os_string(),
-                    std::io::ErrorKind::Other,
-                ))
-            });
-
-        let set_permissions_mock_context = mock_filesystem::set_permissions_context();
-        set_permissions_mock_context.expect().never();
-
-        assert_eq!(
-            args.get_agent_run_directory(),
-            Err(FileSystemError::CreateDirectory(
-                Path::new("/tmp/x").as_os_str().to_os_string(),
-                std::io::ErrorKind::Other
-            ))
-        );
-    }
-
-    // [utest->swdd~agent-prepares-dedicated-run-folder~1]
-    #[test]
-    fn utest_arguments_get_run_directory_use_default_directory_create_permissions_fail() {
-        let _guard = crate::test_helper::MOCKALL_CONTEXT_SYNC.get_lock();
-
-        let args = Arguments {
-            agent_name: "test_agent_name".to_owned(),
-            server_url: DEFAULT_SERVER_ADDRESS.to_string(),
-            run_folder: DEFAULT_RUN_FOLDER.to_owned(),
-            insecure: true,
-            ca_pem: None,
-            crt_pem: None,
-            key_pem: None,
-        };
-
-        let exists_mock_context = mock_filesystem::exists_context();
-        exists_mock_context
-            .expect()
-            .with(predicate::eq(Path::new(DEFAULT_RUN_FOLDER).to_path_buf()))
-            .return_const(false);
-
-        let mk_dir_context = mock_filesystem::make_dir_context();
-        mk_dir_context
-            .expect()
-            .with(predicate::eq(Path::new(DEFAULT_RUN_FOLDER).to_path_buf()))
-            .return_once(|_| Ok(()));
-
-        let set_permissions_mock_context = mock_filesystem::set_permissions_context();
-        set_permissions_mock_context
-            .expect()
-            .with(
-                predicate::eq(Path::new(DEFAULT_RUN_FOLDER).to_path_buf()),
-                predicate::eq(0o777),
-            )
-            .return_once(|_, _| {
-                Err(FileSystemError::Permissions(
-                    Path::new("/tmp/x").as_os_str().to_os_string(),
-                    std::io::ErrorKind::Other,
-                ))
-            });
-
-        assert_eq!(
-            args.get_agent_run_directory(),
-            Err(FileSystemError::Permissions(
-                Path::new("/tmp/x").as_os_str().to_os_string(),
-                std::io::ErrorKind::Other,
-            ))
-        );
-    }
-
-    // [utest->swdd~agent-prepares-dedicated-run-folder~1]
-    #[test]
-    fn utest_arguments_get_run_directory_use_default_directory_exists() {
-        let _guard = crate::test_helper::MOCKALL_CONTEXT_SYNC.get_lock();
-
-        let args = Arguments {
-            agent_name: "test_agent_name".to_owned(),
-            server_url: DEFAULT_SERVER_ADDRESS.to_string(),
-            run_folder: DEFAULT_RUN_FOLDER.to_owned(),
-            insecure: true,
-            ca_pem: None,
-            crt_pem: None,
-            key_pem: None,
-        };
-
-        let exists_mock_context = mock_filesystem::exists_context();
-        exists_mock_context
-            .expect()
-            .with(predicate::eq(Path::new(DEFAULT_RUN_FOLDER).to_path_buf()))
-            .return_const(true);
-
-        let _directory_mock_context =
-            generate_test_directory_mock(DEFAULT_RUN_FOLDER, "test_agent_name_io");
-
-        assert!(args.get_agent_run_directory().is_ok());
-    }
-
-    // [utest->swdd~agent-prepares-dedicated-run-folder~1]
-    #[test]
-    fn utest_arguments_get_run_directory_given_directory_not_found() {
-        let _guard = crate::test_helper::MOCKALL_CONTEXT_SYNC.get_lock();
-
-        let args = Arguments {
-            agent_name: "test_agent_name".to_owned(),
-            server_url: DEFAULT_SERVER_ADDRESS.to_string(),
-            run_folder: "/tmp/x".to_owned(),
-            insecure: true,
-            ca_pem: None,
-            crt_pem: None,
-            key_pem: None,
-        };
-
-        let exists_mock_context = mock_filesystem::exists_context();
-        exists_mock_context
-            .expect()
-            .with(predicate::eq(Path::new("/tmp/x").to_path_buf()))
-            .return_const(false);
-
-        let _directory_mock_context = generate_test_directory_mock("/tmp/x", "test_agent_name_io");
-
-        assert_eq!(
-            args.get_agent_run_directory(),
-            Err(FileSystemError::NotFoundDirectory(
-                Path::new("/tmp/x").as_os_str().to_os_string()
-            ))
-        );
     }
 }
